Split some things into smaller components
