+++ client/src/screens/Feedback/index.js
@@ -1,3 +1,4 @@
+//@flow
 import React, { Component, Fragment } from 'react'
 import { FilterList } from '../../components/FilterList'
 import { Query, Mutation } from 'react-apollo'
@@ -5,7 +6,8 @@ import gql from 'graphql-tag'
 import './styles.scss'
 import { EngagmentSelector } from '../../components/EngagementSelector';
 import { Formik } from 'formik';
-
+import { FeedbackCard } from '../../components/FeedbackCard';
+import { FeedbackFormRender } from './FeedbackFormRender';
 
 
 const FEEDBACK_QUERY = gql`
@@ -28,26 +30,6 @@ const CREATE_FEEDBACK_MUTATION = gql`
   }
 `
 
-
-const formDefinition = {
-  participantFeedback: "ENGAGEMENT",
-  trainerFeedback: [
-    {
-      serviceID: "5cd8dd57a06604360152edd7",
-      engagement: ""
-    },
-    {
-      serviceID: "5cd8dd5da06604360152edd8",
-      engagement: "2"
-    },
-    {
-      serviceID: "5cd8dd65a06604360152edd9",
-      engagement: "1"
-    }
-  ],
-  comment: "this is a test comment"
-}
-
 class FeedbackScreen extends Component {
   constructor(props) {
     super(props)
@@ -66,7 +48,11 @@ class FeedbackScreen extends Component {
   }
 
   handleSelect = (data) => {
-    const { id, value, selectedID } = data
+    const { 
+      id, 
+      value,
+      selectedID
+    } = data
   }
 
   render() {
@@ -76,7 +62,7 @@ class FeedbackScreen extends Component {
           if (loading) return <div>Fetching data...</div>
           if (error) return <div>Error</div>
           const activities = data.activities;
-          const users = data.users; 
+          const users = data.users;
 
           return (
 
@@ -94,91 +80,11 @@ class FeedbackScreen extends Component {
                 }
 
                 return (
-                  <div>
-
-                    <Formik
-                      initialValues={{
-                        comment: "test",
-                      }}
-                      onSubmit={(values, formikBag) => {
-                        console.log(values);
-                        //Values from form come in here. 
-
-                        //NB. there's still a question of matching the correct IDs to the form 
-                        //Below. 
-
-                        //Mutate graphql here. 
-                        feedback({
-                          variables:
-                          {
-                            activityID: values.activity.id,
-                            trainerID: "5cd2cace363cfe4bd9ef981b",
-                            participantID: values.user.id,
-                            participantFeedback: "2",
-                            
-                            //Still need to get the trainer feedback. 
-
-                            comment: values.comment
-                          }
-                        })
-                      }}
-                    >{({
-                      values,
-                      errors,
-                      touched,
-                      handleChange,
-                      handleBlur,
-                      handleSubmit,
-                      isSubmitting,
-
-                      setFieldValue,
-
-                      /* and other goodies */
-                    }) => {
-
-                      const ourHandleChange = (id) => (value) => {
-                        setFieldValue(id, value);
-                      }
-
-
-                      return (
-                        <form onSubmit={handleSubmit}>
-                          <p>Submit Feedback</p>
-                          <div>Activities</div>
-                          <FilterList options={activities.map(v => ({
-                            label: v.name, 
-                            value: v, 
-                          }))} 
-                            handleChange={ourHandleChange('activity')} />
-
-                          <div>Users</div>
-                          <FilterList options={users.map(v=> ({
-                            label: v.email, 
-                            value: v, 
-                          }))} handleChange={ourHandleChange('user')} />
-
-                          <strong>How Engaged Was The Partipant?</strong>
-                          <EngagmentSelector id='0'
-                            handleSelect={ourHandleChange('0')} />
-
-                          <strong>How Much Did They Enjoy it?</strong>
-                          <EngagmentSelector id='1'
-                            handleSelect={ourHandleChange('1')} />
-
-                          <div>Comments</div>
-                          <input
-                            type='text'
-                            id="comment"
-                            value={values.comment}
-                            onChange={handleChange} />
-                          <button
-                            type="submit"
-                          >Submit</button>
-                        </form>
-                      )
-                    }}
-                    </Formik>
-                  </div>
+                  <FeedbackFormRender
+                    feedback={feedback}
+                    activities={activities}
+                    users={users}
+                    initialValues={{}} />
 
                 )
               }}
